Extract the //rewriter/utils package
