+++ rewriter/AttrReader.cc
@@ -6,8 +6,8 @@
 #include "core/Names.h"
 #include "core/core.h"
 #include "core/errors/rewriter.h"
-#include "rewriter/Util.h"
 #include "rewriter/rewriter.h"
+#include "rewriter/utils/Utils.h"
 
 using namespace std;
