+++ rewriter/DSLBuilder.cc
@@ -4,8 +4,8 @@
 #include "core/Context.h"
 #include "core/Names.h"
 #include "core/core.h"
-#include "rewriter/Util.h"
 #include "rewriter/rewriter.h"
+#include "rewriter/utils/Utils.h"
 
 using namespace std;
