+++ rewriter/Delegate.cc
@@ -1,7 +1,7 @@
 #include "rewriter/Delegate.h"
 #include "ast/Helpers.h"
 #include "core/GlobalState.h"
-#include "rewriter/Util.h"
+#include "rewriter/utils/Utils.h"
 #include <optional>
 
 using namespace std;
