+++ rewriter/Prop.cc
@@ -5,7 +5,7 @@
 #include "core/Names.h"
 #include "core/core.h"
 #include "core/errors/rewriter.h"
-#include "rewriter/Util.h"
+#include "rewriter/utils/Utils.h"
 
 using namespace std;
